added function to convert url string to conference object
